apps/bot/requirements.txt
py-clob-client
convex
python-dotenv
loguru
requests
websocket-client
